Elevators/requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==3.0.2
pytest==6.2.5
pytest-flask==1.2.0
Werkzeug==2.2.2